chore(cxx_indexer): pull memcahced-dependent claiming into separate file
